requirements.txt
kafka-python==2.0.2
pyspark==3.5.1
pandas==2.2.2
pyarrow==16.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
apache-airflow-providers-postgres==5.11.1
SQLAlchemy==1.4.52
